Create energyBooster
Energy gels: The season for climbing alpine peaks begins and all climbers stock up on energy gels for climbing. The company offers small and large cuts with 2 pcs. and 5 pcs. energy gels, the price of one gel depends on the fruit from which it is made. Depending on the size of the cut, the price per number of energy gel is different. From the console you can read the fruits, the size of the package (small) 2 pcs. Or (large) 5 pcs.), As well as how many cuts have been ordered. Write a program that calculates the amount to be paid for the order.
Watermelon: small - 56 lv. / big - 28.70 lv.
Mango: small	36.66 lv. - / big - 19.60 lv.
Pineapple: small - 42.10 lv. / big - 24.80 lv.
Raspberry: small - 20 lv. / big - 15.20 lv.
For orders from BGN 400 to BGN 1,000 inclusive there is a 15% discount. For orders over BGN 1,000 there is a 50% discount.