Task: Fitness card
Write a program that checks if the initially available amount is enough to pay for a monthly gym access card. The price of the card depends on the gender of the client and the sport he practices: 
Sports: 
Gym - $ 42 for male; $ 35 for female;
Boxing - $ 41 for male; $ 37 for female;
Yoga - $ 45 for male; $ 42 for female;
Zumba - $ 34 for male;  $ 31 for female;
Dancing - $ 51 for male; $ 53 for female;
Pilates - $ 39 for male; $ 37 for female;
All card prices for students (under 19 years inclusive) have a 20% discount.