Task: Pet food
Ani has two pets - a dog and a cat. Write a program that prepares statistics on pet food for a certain number of days. Every day the dog and the cat eat a different amount of their common food. Every third day they receive a prize - cookies. The amount of cookies is 10% of the total food eaten for the day.
Your program should print statistics on the amount of cookies they ate, what percentage of the original amount of total food they ate, and what percentage of the food the dog ate and how much the cat ate.