Task: Loading suitcases
Write a program that helps you load suitcases in the trunk of an airplane. Each aircraft has a certain trunk capacity. Until you receive the "End" command, you will receive a suitcase volume. The volume of every third suitcase should be increased by 10% due to the loss of space in the way of arrangement. If the free space at a time is less than the volume of the suitcase, loading must be interrupted.