Task: Climb
Georgi decides to break the record for the fastest climb to Mont Blanc. The console introduces the record in seconds that Georgi has to improve, the distance in meters he has to climb and the time in seconds he has to climb 1 meter. Write a program that calculates whether he has coped with the task, given that: the slope of the terrain slows him down every 50 meters by 30 seconds. To calculate the time in seconds for which Georgi will climb the distance to the top and the difference from the record.
When calculating how many times Georgi will slow down as a result of the slope of the field, the result should be rounded down to the nearest whole number.